Fix: Migrations for components and rename `djangocms_frontend.pool` to `djangocms_frontend.component_pool`
