+++ djangocms_frontend/component_pool.py
